Relocate assignment and lecture folder.
