+++ frontend/src/components/teacher/assignment/PopUpMsg.jsx
@@ -1,4 +1,4 @@
-import '../../css/assignment/popUpMsg.css'
+import '../../../css/assignment/popUpMsg.css'
 
 // =====================================================================
 // Pop up message, can be used for "Saved" or "Error" for example
