+++ frontend/src/components/teacher/assignment/chatbot/templates.js
