+++ frontend/src/components/teacher/lecture/LectureCard.jsx
